# functions/src/openapi.yaml
openapi: 3.0.3
info:
  title: CrowdPM API
  version: 1.0.0
  description: >
    API used by CrowdPM to manage devices and retrieve pollution measurements.
  license:
    name: Proprietary
    url: https://crowdpmplatform.web.app
servers:
  - url: https://crowdpmplatform.web.app
    description: Production
  - url: http://127.0.0.1:5001/demo-crowdpm/us-central1/crowdpmApi
    description: Local development (Firebase emulator)
security: []
paths:
...
  /v1/devices:
    get:
      operationId: listDevices
      summary: List devices
      description: >
        Returns the devices the authenticated caller can access. Devices may be
        shared with multiple Firebase users, and the caller only receives entries
        where their UID is a member.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Devices returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        '401':
          description: Authentication required.
    post:
      summary: Register a device
      description: >
        Registers a new device for the authenticated user. The device inherits the
        requesting user's Firebase UID as its owner.
      operationId: registerDevice
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceRequest'
      responses:
        '201':
          description: Device registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceRegistrationResponse'
        '400':
          description: Invalid request payload.
        '401':
          description: Missing or invalid bearer token.
  /v1/devices/{id}/revoke:
    post:
      operationId: revokeDevice
      summary: Revoke a device
      description: >
        Immediately revokes a device owned by the caller. Revocation tears down active access
        tokens and prevents future ingest requests until the node is re-registered.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Device revoked.
        '401':
          description: Missing or invalid bearer token.
        '403':
          description: Caller does not own the device.
        '404':
          description: Device not found.
  /v1/device-activation:
    get:
      operationId: getActivationSession
      summary: Fetch activation session
      description: >
        Returns metadata about a pending pairing session so the user can verify the request
        before approving it. Requires a signed-in Firebase user.
      security:
        - BearerAuth: []
      parameters:
        - name: user_code
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Pending pairing session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivationSession'
        '401':
          description: Missing or invalid bearer token.
        '404':
          description: Session not found.
  /v1/device-activation/authorize:
    post:
      operationId: authorizeActivationSession
      summary: Approve a pending device
      description: >
        Associates a pending pairing session with the caller's account. Accounts with MFA
        enabled must satisfy a fresh second factor before this endpoint succeeds.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_code]
              properties:
                user_code:
                  type: string
      responses:
        '200':
          description: Device authorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivationSession'
        '401':
          description: Missing credentials or MFA requirement not satisfied.
        '404':
          description: Session not found.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    DeviceStatus:
      type: string
      enum:
        - ACTIVE
        - SUSPENDED
    Device:
      type: object
      description: A registered CrowdPM device.
      properties:
        id:
          type: string
          description: Device identifier.
        name:
          type: string
          description: Friendly name supplied by the owner.
        ownerUserId:
          type: string
          description: Primary Firebase UID associated with the device. Legacy field retained for backwards compatibility.
        ownerUserIds:
          type: array
          description: All Firebase UIDs that can access the device.
          items:
            type: string
          minItems: 1
        publicDeviceId:
          type: string
          description: Raw device identifier provided by the owner (before namespacing).
        ownerScope:
          type: string
          description: Slugged segment derived from the owner UID that is prepended to device IDs for uniqueness.
        status:
          $ref: '#/components/schemas/DeviceStatus'
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the device record was created.
        location:
          $ref: '#/components/schemas/Location'
      required:
        - id
        - name
        - ownerUserId
        - status
        - createdAt
    ActivationSession:
      type: object
      description: Metadata about a pending pairing session.
      properties:
        device_code:
          type: string
        user_code:
          type: string
        model:
          type: string
        version:
          type: string
        fingerprint:
          type: string
          description: First bytes of the Ke hash, used for human verification.
        status:
          type: string
          enum: [pending, authorized, redeemed, expired]
        requested_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        requester_ip:
          type: string
          nullable: true
        requester_asn:
          type: string
          nullable: true
        poll_interval:
          type: integer
        authorized_account:
          type: string
          nullable: true
        viewer_account:
          type: string
          nullable: true
      required:
        - device_code
        - user_code
        - model
        - version
        - status
        - requested_at
        - expires_at
    Location:
      type: object
      description: Latitude and longitude in decimal degrees.
      properties:
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
      required:
        - latitude
        - longitude
    RegisterDeviceRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          example: CrowdPM Sensor - Home
          description: Friendly name to display in dashboards.
    DeviceRegistrationResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          description: Newly generated device identifier.
...